# backend/build/swagger.json
{
	"openapi": "3.0.0",
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"UpdateProfileRequest": {
				"properties": {},
				"type": "object",
				"additionalProperties": false
			},
			"TradeDto": {
				"properties": {
					"_id": {
						"type": "string"
					},
					"entryDate": {
						"type": "string",
						"format": "date-time"
					},
					"exitDate": {
						"type": "string",
						"format": "date-time"
					},
					"userId": {
						"type": "string"
					},
					"symbol": {
						"type": "string"
					},
					"entryPrice": {
						"type": "number",
						"format": "double"
					},
					"exitPrice": {
						"type": "number",
						"format": "double"
					},
					"risk": {
						"type": "number",
						"format": "double"
					},
					"reward": {
						"type": "number",
						"format": "double"
					},
					"tags": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"winRate": {
						"type": "number",
						"format": "double"
					},
					"avgProfitLoss": {
						"type": "number",
						"format": "double"
					},
					"riskRewardRatio": {
						"type": "number",
						"format": "double"
					},
					"maxDrawdown": {
						"type": "number",
						"format": "double"
					},
					"profitFactor": {
						"type": "number",
						"format": "double"
					},
					"sharpeRatio": {
						"type": "number",
						"format": "double"
					},
					"volatility": {
						"type": "number",
						"format": "double"
					},
					"sortinoRatio": {
						"type": "number",
						"format": "double"
					},
					"avgHoldingPeriod": {
						"type": "number",
						"format": "double"
					},
					"improvementSuggestions": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"stopLossLevel": {
						"type": "number",
						"format": "double"
					},
					"positionSize": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"_id",
					"entryDate",
					"exitDate",
					"userId",
					"symbol",
					"entryPrice",
					"exitPrice",
					"risk",
					"reward",
					"tags",
					"createdAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ObjectId": {
				"type": "string",
				"description": "A class representation of the BSON ObjectId type."
			},
			"Pick_Document.Exclude_keyofDocument._id__": {
				"properties": {},
				"additionalProperties": {},
				"type": "object",
				"description": "From T, pick a set of properties whose keys are in the union K"
			},
			"Omit_Document._id_": {
				"$ref": "#/components/schemas/Pick_Document.Exclude_keyofDocument._id__",
				"description": "Construct a type with the properties of T except for those in type K."
			},
			"Trade": {
				"properties": {
					"_id": {
						"$ref": "#/components/schemas/ObjectId"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"entry": {
						"type": "number",
						"format": "double"
					},
					"entryQty": {
						"type": "number",
						"format": "double"
					},
					"entryTotal": {
						"type": "number",
						"format": "double"
					},
					"executions": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"exit": {
						"type": "number",
						"format": "double"
					},
					"exitQty": {
						"type": "number",
						"format": "double"
					},
					"exitTotal": {
						"type": "number",
						"format": "double"
					},
					"holdTime": {
						"type": "number",
						"format": "double"
					},
					"lastTransactionAt": {
						"type": "string",
						"format": "date-time"
					},
					"market": {
						"type": "string"
					},
					"openDate": {
						"type": "string",
						"format": "date-time"
					},
					"position": {
						"type": "number",
						"format": "double"
					},
					"positionType": {
						"type": "string"
					},
					"rMultiple": {
						"type": "number",
						"format": "double"
					},
					"urnAmnt": {
						"type": "number",
						"format": "double"
					},
					"returnPercent": {
						"type": "number",
						"format": "double"
					},
					"status": {
						"type": "string"
					},
					"symbol": {
						"type": "string"
					},
					"tags": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"stock": {
						"type": "boolean"
					},
					"crypto": {
						"type": "boolean"
					},
					"forex": {
						"type": "string"
					}
				},
				"required": [
					"_id",
					"createdAt",
					"entry",
					"entryQty",
					"entryTotal",
					"executions",
					"exit",
					"exitQty",
					"exitTotal",
					"holdTime",
					"lastTransactionAt",
					"market",
					"openDate",
					"position",
					"positionType",
					"rMultiple",
					"urnAmnt",
					"returnPercent",
					"status",
					"symbol",
					"tags"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {
			"jwt": {
				"type": "jwt",
				"name": "token",
				"in": "headers"
			},
			"tsoa_auth": {
				"type": "oauth2",
				"flows": {
					"implicit": {
						"authorizationUrl": "http://swagger.io/api/oauth/dialog",
						"scopes": {}
					}
				}
			}
		}
	},
	"info": {
		"title": "backend",
		"version": "1.0.0",
		"license": {
			"name": "MIT"
		},
		"contact": {
			"name": "Bartek-Figat"
		}
	},
	"paths": {
		"/profile/user-profile-info": {
			"get": {
				"operationId": "GetUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"data": {
											"type": "string"
										}
									},
									"required": [
										"data"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": []
			}
		},
		"/profile/update-profile": {
			"post": {
				"operationId": "UpdateProfile",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"data": {
											"type": "string"
										}
									},
									"required": [
										"data"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateProfileRequest"
							}
						}
					}
				}
			}
		},
		"/profile/all-users": {
			"get": {
				"operationId": "GetAllUsers",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"data": {
											"type": "string"
										}
									},
									"required": [
										"data"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": []
			}
		},
		"/profile/delete-user/{userId}": {
			"delete": {
				"operationId": "DeleteUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"data": {
											"type": "string"
										}
									},
									"required": [
										"data"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/profile/user/{userId}": {
			"get": {
				"operationId": "GetUserById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"data": {
											"type": "string"
										}
									},
									"required": [
										"data"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/custom-trades/get-trade/{tradeId}": {
			"get": {
				"operationId": "GetTradeById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/TradeDto"
								}
							}
						}
					}
				},
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "tradeId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/custom-trades/get-all-trades": {
			"get": {
				"operationId": "GetAllTrades",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/TradeDto"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": []
			}
		},
		"/custom-trades/create": {
			"post": {
				"operationId": "CreateTrade",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"tradeType": {
											"type": "string",
											"enum": [
												"stock",
												"forex",
												"crypto"
											]
										},
										"positionSize": {
											"type": "number",
											"format": "double"
										},
										"stopLossLevel": {
											"type": "number",
											"format": "double"
										},
										"createdAt": {
											"type": "string",
											"format": "date-time"
										},
										"tags": {
											"items": {
												"type": "string"
											},
											"type": "array"
										},
										"reward": {
											"type": "number",
											"format": "double"
										},
										"risk": {
											"type": "number",
											"format": "double"
										},
										"exitPrice": {
											"type": "number",
											"format": "double"
										},
										"entryPrice": {
											"type": "number",
											"format": "double"
										},
										"symbol": {
											"type": "string"
										},
										"exitDate": {
											"type": "string",
											"format": "date-time"
										},
										"entryDate": {
											"type": "string",
											"format": "date-time"
										}
									},
									"required": [
										"tradeType",
										"createdAt",
										"tags",
										"reward",
										"risk",
										"exitPrice",
										"entryPrice",
										"symbol",
										"exitDate",
										"entryDate"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {}
						}
					}
				}
			}
		},
		"/custom-trades/update-existing-trade/{tradeId}": {
			"put": {
				"operationId": "UpdateTrade",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/TradeDto"
								}
							}
						}
					}
				},
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "tradeId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Trade"
							}
						}
					}
				}
			}
		},
		"/custom-trades/filter-trades": {
			"get": {
				"operationId": "FilterTrades",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/TradeDto"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"default": 1,
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"default": 10,
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/custom-trades/company/{symbol}": {
			"get": {
				"operationId": "GetCompanyProfile",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "symbol",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/auth/register": {
			"post": {
				"operationId": "Registration",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"tags": [
					"Auth"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {}
						}
					}
				}
			}
		},
		"/auth/login": {
			"post": {
				"operationId": "Login",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"token": {
											"type": "string"
										}
									},
									"required": [
										"token"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Auth"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {}
						}
					}
				}
			}
		},
		"/auth/logout": {
			"get": {
				"operationId": "Logout",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"tags": [
					"Auth"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": []
			}
		},
		"/auth/logout-all": {
			"get": {
				"operationId": "LogoutFromAllDevices",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"tags": [
					"Auth"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": []
			}
		},
		"/auth/confirm-email/{authToken}": {
			"get": {
				"operationId": "EmailConfirmation",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"tags": [
					"Auth"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "authToken",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}